separated function to hook
