+++ src/components/web3Connection/Account/Account.tsx
@@ -5,13 +5,10 @@ import {
   useBreakpointValue,
   useDisclosure,
 } from "@chakra-ui/react"
-import { BigNumber } from "@ethersproject/bignumber"
-import { ExternalProvider } from "@ethersproject/providers"
 import { UnsupportedChainIdError, useWeb3React } from "@web3-react/core"
 import Card from "components/common/Card"
 import { useCommunity } from "components/community/Context"
 import { Web3Connection } from "components/web3Connection/Web3ConnectionManager"
-import { addToWalletData, Chains } from "connectors"
 import useBalance from "hooks/useBalance"
 import { LinkBreak, SignIn, Wallet } from "phosphor-react"
 import { useContext } from "react"
@@ -19,6 +16,7 @@ import type { Token } from "temporaryData/types"
 import shortenHex from "utils/shortenHex"
 import AccountModal from "../AccountModal"
 import useENSName from "./hooks/useENSName"
+import useRequestNetworkChange from "./hooks/useRequestNetworkChange"
 
 type Props = {
   token: Token
@@ -31,40 +29,7 @@ const Account = (): JSX.Element => {
   const ENSName = useENSName(account)
   const { isOpen, onOpen, onClose } = useDisclosure()
   const shortenHexText = useBreakpointValue({ base: 2, sm: 3 })
-
-  const requestNetworkChange = async () => {
-    const { ethereum } = window as Window &
-      typeof globalThis & { ethereum: ExternalProvider }
-
-    // If we are on the all communities page, we dont have a chainData from communityData, setting Polygon as a default there
-    const chainName = communityData?.chainData.name ?? "polygon"
-
-    try {
-      await ethereum.request({
-        method: "wallet_switchEthereumChain",
-        params: [
-          {
-            chainId: BigNumber.from(Chains[chainName]).toHexString(),
-          },
-        ],
-      })
-    } catch (e) {
-      // This error code indicates that the chain has not been added to MetaMask.
-      if (e.code === 4902) {
-        try {
-          if (!addToWalletData[chainName]) throw Error()
-
-          await ethereum.request({
-            method: "wallet_addEthereumChain",
-            params: [addToWalletData[chainName]],
-          })
-        } catch (addError) {
-          console.error("Failed to add network to MetaMask")
-        }
-      }
-      // handle other "switch" errors
-    }
-  }
+  const requestNetworkChange = useRequestNetworkChange()
 
   if (typeof window === "undefined") {
     return (
